+++ packages/cloud_firestore/cloud_firestore_web/lib/cloud_firestore_web.dart
@@ -3,12 +3,12 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:cloud_firestore_platform_interface/cloud_firestore_platform_interface.dart';
+import 'package:cloud_firestore_web/src/internals.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_core_web/firebase_core_web_interop.dart'
     as core_interop;
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
 
-import 'src/utils/web_utils.dart';
 import 'src/collection_reference_web.dart';
 import 'src/field_value_factory_web.dart';
 import 'src/document_reference_web.dart';