+++ packages/cloud_firestore/cloud_firestore_web/lib/src/transaction_web.dart
@@ -4,6 +4,7 @@
 
 import 'package:cloud_firestore_platform_interface/cloud_firestore_platform_interface.dart';
 
+import 'internals.dart';
 import 'interop/firestore.dart' as firestore_interop;
 import 'utils/codec_utility.dart';
 import 'utils/web_utils.dart';
@@ -27,21 +28,27 @@ class TransactionWeb extends TransactionPlatform {
   }
 
   @override
-  Future<DocumentSnapshotPlatform> get(String documentPath) async {
-    final webDocumentSnapshot = await guard(() {
-      return _webTransactionDelegate
-          .get(_webFirestoreDelegate.doc(documentPath));
-    });
-
-    return guardSync(
-        () => convertWebDocumentSnapshot(_firestore, webDocumentSnapshot));
+  Future<DocumentSnapshotPlatform> get(String documentPath) {
+    return guard(
+      () async {
+        final webDocumentSnapshot = await _webTransactionDelegate
+            .get(_webFirestoreDelegate.doc(documentPath));
+        return convertWebDocumentSnapshot(_firestore, webDocumentSnapshot);
+      },
+    );
   }
 
   @override
-  TransactionWeb set(String documentPath, Map<String, dynamic> data,
-      [SetOptions? options]) {
-    _webTransactionDelegate.set(_webFirestoreDelegate.doc(documentPath),
-        CodecUtility.encodeMapData(data)!, convertSetOptions(options));
+  TransactionWeb set(
+    String documentPath,
+    Map<String, dynamic> data, [
+    SetOptions? options,
+  ]) {
+    _webTransactionDelegate.set(
+      _webFirestoreDelegate.doc(documentPath),
+      CodecUtility.encodeMapData(data)!,
+      convertSetOptions(options),
+    );
     return this;
   }
 
@@ -50,8 +57,10 @@ class TransactionWeb extends TransactionPlatform {
     String documentPath,
     Map<String, dynamic> data,
   ) {
-    _webTransactionDelegate.update(_webFirestoreDelegate.doc(documentPath),
-        CodecUtility.encodeMapData(data)!);
+    _webTransactionDelegate.update(
+      _webFirestoreDelegate.doc(documentPath),
+      CodecUtility.encodeMapData(data)!,
+    );
     return this;
   }
 }