+++ packages/firebase_auth/firebase_auth_platform_interface/test/method_channel_tests/method_channel_firebase_auth_test.dart
@@ -7,9 +7,10 @@ import 'dart:async';
 import 'package:firebase_auth_platform_interface/firebase_auth_platform_interface.dart';
 import 'package:firebase_auth_platform_interface/src/method_channel/method_channel_firebase_auth.dart';
 import 'package:firebase_auth_platform_interface/src/method_channel/method_channel_user.dart';
+import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:firebase_core/firebase_core.dart';
+
 import '../mock.dart';
 
 void main() {
@@ -965,35 +966,35 @@ void main() {
       });
 
       test('listens to incoming changes', () async {
-        const String testEmail = 'testauthstate@email.com';
-        Stream<UserPlatform?> stream = auth.authStateChanges();
-        int call = 0;
+        Stream<UserPlatform?> stream =
+            auth.authStateChanges().asBroadcastStream();
 
-        subscription = stream.listen(
-          expectAsync1((UserPlatform? user) {
-            call++;
-            if (call == 1) {
-              expect(user, isA<UserPlatform>());
-              expect(user!.email, isNull);
-            } else if (call == 2) {
-              expect(user!.email, equals(testEmail));
-            } else {
-              fail('Should not have been called');
-            }
-          }, count: 2, reason: 'Stream should only have been called 2 times'),
-        );
+        await expectLater(stream, emits(isNull));
+        expect(auth.currentUser, equals(isNull));
 
         await injectEventChannelResponse('authStateChannel', {'user': user});
 
-        final Map<String, dynamic> updatedUser = <String, dynamic>{
-          'email': testEmail,
-        };
-        await injectEventChannelResponse(
-          'authStateChannel',
-          {'user': updatedUser},
+        await expectLater(
+          stream,
+          emits(isA<UserPlatform>().having((e) => e.uid, 'uid', kMockUid)),
+        );
+        expect(auth.currentUser!.uid, equals(kMockUid));
+
+        expect(log, isEmpty);
+      });
+
+      test('emits the latest user available', () async {
+        Stream<UserPlatform?> stream = auth.authStateChanges();
+
+        await injectEventChannelResponse('authStateChannel', {'user': user});
+
+        await expectLater(
+          stream,
+          emits(isA<UserPlatform>().having((e) => e.uid, 'uid', kMockUid)),
         );
 
-        expect(log, equals([]));
+        expect(auth.currentUser!.uid, equals(kMockUid));
+        expect(log, isEmpty);
       });
     });
 
@@ -1010,28 +1011,36 @@ void main() {
       });
 
       test('listens to incoming changes', () async {
-        Stream<UserPlatform?> stream = auth.idTokenChanges();
-        int call = 0;
+        Stream<UserPlatform?> stream =
+            auth.idTokenChanges().asBroadcastStream();
 
-        subscription = stream.listen(
-          expectAsync1((UserPlatform? user) {
-            call++;
-            if (call == 1) {
-              expect(user, isNull);
-            } else if (call == 2) {
-              expect(user!.uid, isA<String>());
-              expect(user.uid, equals(kMockUid));
-              expect(auth.currentUser!.uid, equals(user.uid));
-            } else {
-              fail('Should not have been called');
-            }
-          }, count: 2, reason: 'Stream should only have been called 2 times'),
+        await expectLater(stream, emits(isNull));
+        expect(auth.currentUser, equals(isNull));
+
+        await injectEventChannelResponse('idTokenChannel', {'user': user});
+
+        await expectLater(
+          stream,
+          emits(isA<UserPlatform>().having((e) => e.uid, 'uid', kMockUid)),
         );
+        expect(auth.currentUser!.uid, equals(kMockUid));
 
+        expect(log, isEmpty);
+      });
+
+      test('emits the latest user available', () async {
+        Stream<UserPlatform?> stream = auth.idTokenChanges();
+        
         await injectEventChannelResponse('idTokenChannel', {'user': null});
         await injectEventChannelResponse('idTokenChannel', {'user': user});
 
-        expect(log, equals([]));
+        await expectLater(
+          stream,
+          emits(isA<UserPlatform>().having((e) => e.uid, 'uid', kMockUid)),
+        );
+
+        expect(auth.currentUser!.uid, equals(kMockUid));
+        expect(log, isEmpty);
       });
     });
 
@@ -1048,31 +1057,36 @@ void main() {
       });
 
       test('listens to incoming changes', () async {
-        Stream<UserPlatform?> stream = auth.userChanges();
-        int call = 0;
+        Stream<UserPlatform?> stream = auth.userChanges().asBroadcastStream();
 
-        subscription = stream.listen(
-          expectAsync1((UserPlatform? user) {
-            call++;
-            if (call == 1) {
-              expect(user, isNull);
-              expect(auth.currentUser, equals(isNull));
-            } else if (call == 2) {
-              expect(user!.uid, isA<String>());
-              expect(user.uid, equals(kMockUid));
-              expect(auth.currentUser!.uid, equals(user.uid));
-            } else {
-              fail('Should not have been called');
-            }
-          }, count: 2, reason: 'Stream should only have been called 2 times'),
+        await expectLater(stream, emits(isNull));
+        expect(auth.currentUser, equals(isNull));
+
+        await injectEventChannelResponse('idTokenChannel', {'user': user});
+
+        await expectLater(
+          stream,
+          emits(isA<UserPlatform>().having((e) => e.uid, 'uid', kMockUid)),
         );
+        expect(auth.currentUser!.uid, equals(kMockUid));
 
+        expect(log, isEmpty);
+      });
+
+      test('emits the latest user available', () async {
+        Stream<UserPlatform?> stream = auth.userChanges();
         // id token change events will trigger setCurrentUser()
         // and hence userChange events
         await injectEventChannelResponse('idTokenChannel', {'user': null});
         await injectEventChannelResponse('idTokenChannel', {'user': user});
 
-        expect(log, equals([]));
+        await expectLater(
+          stream,
+          emits(isA<UserPlatform>().having((e) => e.uid, 'uid', kMockUid)),
+        );
+
+        expect(auth.currentUser!.uid, equals(kMockUid));
+        expect(log, isEmpty);
       });
     });
   });