+++ packages/firebase_remote_config/firebase_remote_config/example/test_driver/firebase_remote_config_e2e_test.dart
@@ -2,6 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+// ignore: import_of_legacy_library_into_null_safe
 import 'package:integration_test/integration_test_driver.dart';
 
 Future<void> main() => integrationDriver();