+++ packages/firebase_core/firebase_core/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.4
+
+ - **REFACTOR**: Share guard functions accross plugins (#5783).
+ - **CHORE**: update Web plugins to use Firebase JS SDK version 8.4.1 (#4464).
+
 ## 1.0.3
 
  - **REFACTOR**: Remove deprecated gradle property android.enableR8=true (#5463).